feat(core): add video file support, extend DB schema, and overhaul UI assets
- Updated .gitignore to exclude *.mp4 video files.
- Extended database migration logic and added new columns/models for email marketing, notebooks, and tasks.
- Refactored backend models (email_marketing, notebook, task) with schema improvements.
- Added numerous admin and discussion router endpoints and enhanced task manager utilities.
- Integrated a large set of compiled frontend UI assets and updated Vue components (AdminPanel, EmailMarketing, layout, modals, views).
- Updated main entry point and requirements to reflect new dependencies.

+++ backend/models/task.py
@@ -14,7 +14,7 @@ class TaskInfo(BaseModel):
     name: str
     description: Optional[str] = None
     status: TaskStatus
-    progress: int
+    progress: float
     logs: List[TaskLogMessage]
     result: Optional[Any] = None
     error: Optional[str] = None
